+++ realm/realm-library/src/main/java/io/realm/internal/Table.java
@@ -37,12 +37,13 @@ public class Table implements TableSchema, NativeObject {
     public static final int NO_MATCH = -1;
 
     private static final String TABLE_PREFIX = Util.getTablePrefix();
-    private static final String PRIMARY_KEY_TABLE_NAME = "pk";
+    static final String PRIMARY_KEY_TABLE_NAME = "pk";
     private static final String PRIMARY_KEY_CLASS_COLUMN_NAME = "pk_table";
     private static final long PRIMARY_KEY_CLASS_COLUMN_INDEX = 0;
     private static final String PRIMARY_KEY_FIELD_COLUMN_NAME = "pk_property";
     private static final long PRIMARY_KEY_FIELD_COLUMN_INDEX = 1;
     public static final long NO_PRIMARY_KEY = -2;
+    public static final String OBJECT_ID_COLUMN_NAME = "__OID"; // Column for stable ID
 
     private static final long nativeFinalizerPtr = nativeGetFinalizerPtr();
 
@@ -610,7 +611,7 @@ public class Table implements TableSchema, NativeObject {
 
         // FIXME: The PK table creation should be handle by Object Store after integration of OS Schema.
         if (!sharedRealm.hasTable(PRIMARY_KEY_TABLE_NAME)) {
-            sharedRealm.createTable(PRIMARY_KEY_TABLE_NAME);
+            sharedRealm.createPkTable();
         }
 
         Table pkTable = sharedRealm.getTable(PRIMARY_KEY_TABLE_NAME);